requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.2
celery==5.3.1
redis==4.6.0
azure-devops==6.0.0b4
pydantic==2.4.2
httpx==0.25.0
aiohttp==3.8.5
tenacity==8.2.3
prometheus-client==0.17.1
structlog==23.2.0
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
responses==0.24.1
pytest-cov==4.1.0
markupsafe==2.1.3
